pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exargs"
version = "0.1.0"
description = "A configurable resolver supporting variables and dependency analysis"
authors = [
  {name = "Qiaoy", email = "qiaoy@buaa.edu.cn"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
dependencies = [
  "pyyaml"
]

[project.urls]
Homepage = "https://github.com/yourusername/exargs"

[tool.setuptools]
packages = ["exargs"]
